remove Windows support
