+++ src/tests/CMakeLists.txt
@@ -28,21 +28,16 @@
 # (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-
-if (NOT WIN32)
-		pkg_check_modules(Libndctl QUIET libndctl)
-		pkg_check_modules(Libdaxctl QUIET libdaxctl)
-		if (Libndctl_FOUND AND Libdaxctl_FOUND)
-			include(${CMAKE_CURRENT_LIST_DIR}/ras/CMakeLists.txt)
-		else ()
-			if (NOT Libndctl_FOUND AND NOT Libdaxctl_FOUND)
-				message(WARNING "Cannot find libndctl and libdaxctl libraries. Skip building RAS tests.")
-			elseif (NOT Libndctl_FOUND)
-				message(WARNING "Cannot find libndctl library. Skip building RAS tests.")
-			else ()
-				message(WARNING "Cannot find libdaxctl library. Skip building RAS tests.")
-			endif()
-		endif ()
+pkg_check_modules(Libndctl QUIET libndctl)
+pkg_check_modules(Libdaxctl QUIET libdaxctl)
+if (Libndctl_FOUND AND Libdaxctl_FOUND)
+	include(${CMAKE_CURRENT_LIST_DIR}/ras/CMakeLists.txt)
 else ()
-		include(${CMAKE_CURRENT_LIST_DIR}/ras/CMakeLists.txt)
-endif()
+	if (NOT Libndctl_FOUND AND NOT Libdaxctl_FOUND)
+		message(WARNING "Cannot find libndctl and libdaxctl libraries. Skip building RAS tests.")
+	elseif (NOT Libndctl_FOUND)
+		message(WARNING "Cannot find libndctl library. Skip building RAS tests.")
+	else ()
+		message(WARNING "Cannot find libdaxctl library. Skip building RAS tests.")
+	endif()
+endif ()
