+++ src/tests/ras/utils/CMakeLists.txt
@@ -41,16 +41,11 @@ set_source_groups("${PREFIX_FILTER}" ${ras_utils_SRC})
 add_library(RasUtils STATIC ${ras_utils_SRC})
 add_dependencies(RasUtils Utils)
 
-if (NOT WIN32)
-	pkg_check_modules(Libndctl REQUIRED libndctl)
-	include_directories(${Libndctl_INCLUDE_DIRS})
-	link_directories(${Libndctl_LIBRARY_DIRS})
-	pkg_check_modules(Libdaxctl REQUIRED libdaxctl)
-	include_directories(${Libdaxctl_INCLUDE_DIRS})
-	link_directories(${Libdaxctl_LIBRARY_DIRS})	
-	target_link_libraries(RasUtils Utils ${Libpmemobj_LIBRARIES}
-            ${Libdaxctl_LIBRARIES} ${Libndctl_LIBRARIES})
-else()
-	target_link_libraries(RasUtils Utils ${Libpmemobj_LIBRARIES})
-endif()
-
+pkg_check_modules(Libndctl REQUIRED libndctl)
+include_directories(${Libndctl_INCLUDE_DIRS})
+link_directories(${Libndctl_LIBRARY_DIRS})
+pkg_check_modules(Libdaxctl REQUIRED libdaxctl)
+include_directories(${Libdaxctl_INCLUDE_DIRS})
+link_directories(${Libdaxctl_LIBRARY_DIRS})
+target_link_libraries(RasUtils Utils ${Libpmemobj_LIBRARIES}
+        ${Libdaxctl_LIBRARIES} ${Libndctl_LIBRARIES})
