+++ src/tests/ras/utils/dimm/dimm.cc
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018, Intel Corporation
+ * Copyright 2018-2023, Intel Corporation
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -29,8 +29,6 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-
-#ifdef __linux__
 
 #include <linux/limits.h>
 #include <cstring>
@@ -188,5 +186,3 @@ DimmNamespace::~DimmNamespace() {
 
   ctx_ = nullptr;
 }
-
-#endif  // __linux__
