+++ src/tests/ras/utils/inject_manager/inject_manager.cc
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018, Intel Corporation
+ * Copyright 2018-2023, Intel Corporation
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -29,8 +29,6 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-
-#ifdef __linux__
 
 #include "inject_manager.h"
 
@@ -172,5 +170,3 @@ bool InjectManager::IsLastShutdownSafe(
   }
   return true;
 }
-
-#endif  // __LINUX__
