+++ src/utils/constants.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017, Intel Corporation
+ * Copyright 2017-2023, Intel Corporation
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -36,19 +36,9 @@
 #include <map>
 #include <string>
 
-#ifdef _WIN32
-/* On Windows there is no permission for execution. Also windows doesn't
-* support groups of users in UNIX way */
-const int PERMISSION_MASK = 0600;
-#else
 const int PERMISSION_MASK = 0777;
-#endif  // _WIN32
 
-#ifdef _WIN32
-const std::string SEPARATOR = "\\";
-#else
 const std::string SEPARATOR = "/";
-#endif  // _WIN32
 
 static const size_t KIBIBYTE = 1 << 10;
 static const size_t MEBIBYTE = KIBIBYTE << 10;
