+++ src/utils/shell/i_shell.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017-2018, Intel Corporation
+ * Copyright 2017-2023, Intel Corporation
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -44,11 +44,6 @@
 
 const int BUFFER_SIZE = 128;
 
-#ifdef _WIN32
-#define popen _popen
-#define pclose _pclose
-#endif  // _WIN32
-
 struct PipeDeleter {
   void operator()(FILE *pipe) {
     if (pipe) {
@@ -66,20 +61,12 @@ class IShell : NonCopyable {
   Output<char> output_;
   bool print_log_ = false;
 
-#ifndef _WIN32
   std::string address_;
-#endif
-
-#ifdef _WIN32
-  Output<wchar_t> w_output_;
-#endif
 
  public:
   IShell(){};
   IShell(bool print_log) : print_log_(print_log){};
-#ifndef _WIN32
   IShell(const std::string &address) : address_(address){};
-#endif
   /*
    * GetLastOutput -- returns last received output.
    */
@@ -92,20 +79,6 @@ class IShell : NonCopyable {
    */
   Output<char> ExecuteCommand(const std::string &cmd);
 
-#ifdef _WIN32
-  /*
-   * GetLastOutput -- returns last received output.
-   */
-  Output<wchar_t> GetWLastOutput() const {
-    return w_output_;
-  }
-
-  /*
-   * ExecuteCommand -- performs command by creating pipe with read mode. Returns
-   * Output object on success, throws std::exception otherwise.
-   */
-  Output<wchar_t> ExecuteCommand(const std::wstring &cmd);
-#endif  // _WIN32
 };
 
 #endif  // !PMDK_TESTS_SRC_UTILS_SHELL_I_SHELL_H_
