--- src/utils/shell/windows_shell.cc
@@ -1,64 +1,0 @@
-/*
- * Copyright 2018, Intel Corporation
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- *
- *     * Redistributions of source code must retain the above copyright
- *       notice, this list of conditions and the following disclaimer.
- *
- *     * Redistributions in binary form must reproduce the above copyright
- *       notice, this list of conditions and the following disclaimer in
- *       the documentation and/or other materials provided with the
- *       distribution.
- *
- *     * Neither the name of the copyright holder nor the names of its
- *       contributors may be used to endorse or promote products derived
- *       from this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-#ifdef _WIN32
-
-#include "i_shell.h"
-
-Output<wchar_t> IShell::ExecuteCommand(const std::wstring &cmd) {
-  std::wstring command = L"PowerShell -Command " + cmd + L" 2>&1";
-
-  std::unique_ptr<FILE, PipeDeleter> pipe(_wpopen(command.c_str(), L"r"));
-
-  if (!pipe) {
-    throw std::exception("popen failed");
-  }
-
-  wchar_t buffer[BUFFER_SIZE];
-  std::wstring out_buffer;
-
-  while (fgetws(buffer, BUFFER_SIZE, pipe.get())) {
-    out_buffer.append(buffer);
-  }
-
-  auto s_pipe = pipe.release();
-  int exit_code = pclose(s_pipe);
-  w_output_ = Output<wchar_t>(exit_code, out_buffer);
-
-  if (print_log_) {
-    std::wcout << out_buffer << std::endl;
-  }
-
-  return w_output_;
-}
-
-#endif  // _WIN32
